# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'PlayerApi'","description":"'PlayerApi'","contact":{"name":"Imanol Suppo Alaniz","email":"114184@frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'PlayerApi'"}],"paths":{"/plays_rps":{"post":{"tags":["play-rps-controller"],"operationId":"savePlay","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayRps"}}}}}}},"/players":{"post":{"tags":["player-controller"],"summary":"Create a new Player","description":"Return the player created with your id","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/matches":{"post":{"tags":["match-controller"],"operationId":"saveMatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/MatchRps"}]}}}}}}},"/matches/{id}/plays":{"post":{"tags":["match-controller"],"operationId":"saveMatch_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/PlayRpsDTO"}]}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Play"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/plays_rps/{id}":{"get":{"tags":["play-rps-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayRps"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/players/{id}/matches":{"get":{"tags":["player-controller"],"operationId":"getMatchesOfPlayer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/MatchRps"}]}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it´s alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/matches/{id}":{"get":{"tags":["match-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"oneOf":[{"$ref":"#/components/schemas/MatchRps"}]}}}}}}}},"components":{"schemas":{"PlayDTO":{"type":"object","properties":{"matchRpsId":{"type":"integer","format":"int64"},"shapeHandPlayer1":{"type":"string","enum":["ROCK","PAPER","SCISSORS"]}}},"PlayRps":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"matchRpsId":{"type":"integer","format":"int64"},"shapeHandPlayer1":{"type":"string","enum":["ROCK","PAPER","SCISSORS"]},"shapeHandPlayer2":{"type":"string","enum":["ROCK","PAPER","SCISSORS"]},"winnerId":{"type":"integer","format":"int64"}}},"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"the player id","format":"int64","example":1},"userName":{"title":"Player user name","type":"string","description":"the player user name","example":"Imanol"},"password":{"title":"Player password","type":"string","description":"the player password","example":"Password123"},"email":{"title":"Player email","type":"string","description":"the player email","example":"email@email.com"},"avatar":{"title":"Player avatar url","type":"string","description":"the player avatar ","nullable":true,"example":"https://localhost:8080/avatars/myUserName"},"lastLoginDate":{"title":"Player last login","type":"string","description":"the player last login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"MatchDTO":{"required":["gameId","playerId"],"type":"object","properties":{"gameId":{"type":"integer","format":"int64"},"playerId":{"type":"integer","format":"int64"}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"rules":{"type":"string"}}},"Match":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"game":{"$ref":"#/components/schemas/Game"},"player1":{"$ref":"#/components/schemas/Player"},"player2":{"$ref":"#/components/schemas/Player"},"createdDate":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["STARTED","FINISHED","CANCELED"]}}},"MatchRps":{"type":"object","allOf":[{"$ref":"#/components/schemas/Match"},{"type":"object","properties":{"numberOfPlays":{"type":"integer","format":"int32"},"remainderPlays":{"type":"integer","format":"int32"},"player1Score":{"type":"integer","format":"int32"},"player2Score":{"type":"integer","format":"int32"},"plays":{"type":"array","items":{"$ref":"#/components/schemas/PlayRps"}},"winnerId":{"type":"integer","format":"int64"}}}]},"PlayRequest":{"type":"object"},"PlayRpsDTO":{"required":["shape_hand_player_1"],"type":"object","allOf":[{"$ref":"#/components/schemas/PlayRequest"},{"type":"object","properties":{"shape_hand_player_1":{"type":"string","enum":["ROCK","PAPER","SCISSORS"]},"shape_hand_player_2":{"type":"string","enum":["ROCK","PAPER","SCISSORS"]}}}]},"Play":{"type":"object"},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email","type":"string","description":"The player email","example":"email@email.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity used to logged in","type":"string","description":"The type of identity provided to logged in","example":"USERNAME or EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"Username","type":"string","description":"The player username","example":"myUserName"}}}]},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"title":"Email or UserName to logged in","type":"string","description":"The player email or username","example":"email@gmail.com or username"}}}}}}