# src/main/java/ar/edu/utn/frc/tup/lciii/playerapi/dtos/play/PlayRequest.java
package ar.edu.utn.frc.tup.lciii.playerapi.dtos.play;


import com.fasterxml.jackson.annotation.JsonSubTypes;
import com.fasterxml.jackson.annotation.JsonTypeInfo;

@JsonTypeInfo(use = JsonTypeInfo.Id.DEDUCTION)
@JsonSubTypes({
        @JsonSubTypes.Type(value = PlayRpsDTO.class)
})
public interface PlayRequest {
}