# src/main/java/ar/edu/utn/frc/tup/lciii/playerapi/models/Game.java
package ar.edu.utn.frc.tup.lciii.playerapi.models;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Game {
    private Long id;
    private String code;
    private String name;
    private String description;
    private String rules;
}